Imprve the API for starting hosts

+++ SwiftLSPClient/LanguageServerProcessHost.swift
@@ -12,11 +12,13 @@ public class LanguageServerProcessHost {
     private let process: Process
     private let transport: StdioDataTransport
     private let server: LanguageServer
+    private var launched: Bool
 
     public init(path: String, arguments: [String]) {
         self.process = Process()
         self.transport = StdioDataTransport()
         self.server = JSONRPCLanguageServer(dataTransport: transport)
+        self.launched = false
 
         process.standardInput = transport.stdinPipe
         process.standardOutput = transport.stdoutPipe
@@ -32,13 +34,32 @@ public class LanguageServerProcessHost {
 
     public func start(block: @escaping (LanguageServer?) -> Void) {
         process.launch()
+        launched = true
 
         block(server)
     }
 
     public func stop(block: @escaping () -> Void) {
-        process.terminate()
+        if launched {
+            process.terminate()
+        }
 
         block()
+    }
+
+    public func getServer(_ block: @escaping (LanguageServer?) -> Void) {
+        switch (process.isRunning, launched) {
+        case (true, true):
+            block(server)
+        case (true, false):
+            // nonsense state
+            block(nil)
+        case (false, true):
+            block(nil)
+        case (false, false):
+            start { (s) in
+                block(s)
+            }
+        }
     }
 }
